Edge Point detection -> Three Line detection
The Edge Point will be violated by the noise. 
Adding the dynamic first line when it has detected above the threshold color pixels, then the second and the third line would be based on the first line by adding 20 and 40.
Averaging how many times had we counted the shape every 4000 frames. By doing these, we would have a more constant result. And the  LED won't keep blinking all the time.
NOW, I am using a fix line number when detecting the color pixel numbers. It is much stable than the dynamic line detection. 
The advantage of the dynamic line detection method is benefit when the shape is higher or lower than the center. It should be more robust when the shape isn't centered.